src/python/requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
eland==9.0.1
elastic-transport==9.1.0
elasticsearch==9.1.0
executing==2.2.0
fastparquet==2024.11.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lightgbm==4.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.5.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xgboost==3.0.3
